update cli to start connector by default (#22)
